requirements.txt
dnspython==2.4.2
pymongo==4.6.0
python-dotenv==1.0.0
